posts/signal/ecg-library-comparison/requirements.txt
biosppy==1.0.0
click==8.0.4
heartpy==1.2.7
joblib==1.1.0
matplotlib==3.5.1
neurokit2==0.2.2
numpy==1.22.2
pandas==2.0.1
py_ecg_detectors==1.3.2
seaborn==0.11.2
sleepecg==0.5.3
wfdb==4.1.1
